Digital Elections | Kafka | PySpark | Redis | Streamlit/app/requirements.txt
fastapi
uvicorn
confluent-kafka
redis
psycopg2-binary
faker
pydantic

